webhook/requirements.txt
# Core Cloud Function dependencies
functions-framework>=3.4.0
flask>=2.3.0

# Google Cloud dependencies
google-cloud-bigquery>=3.10.0
google-cloud-storage>=2.10.0
google-cloud-logging>=3.5.0
google-auth>=2.20.0

# HTTP client for API calls
requests>=2.31.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
